--- src/main/java/chess/dto/ErrorResponseDto.java
@@ -1,10 +1,0 @@
-package chess.dto;
-
-public class ErrorResponseDto {
-
-	private final String errorMessage;
-
-	public ErrorResponseDto(final String errorMessage) {
-		this.errorMessage = errorMessage;
-	}
-}
