+++ src/main/java/chess/dto/response/ErrorDto.java
@@ -1,0 +1,10 @@
+package chess.dto.response;
+
+public class ErrorDto {
+
+	private final String errorMessage;
+
+	public ErrorDto(final String errorMessage) {
+		this.errorMessage = errorMessage;
+	}
+}
