+++ src/main/java/chess/dto/response/GamesDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.response;
 
 import chess.domain.game.Game;
 import java.util.ArrayList;
