+++ src/main/java/chess/dto/response/WinnerDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.response;
 
 public class WinnerDto {
