+++ src/test/java/chess/service/ChessServiceTest.java
@@ -6,10 +6,10 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import chess.controller.PositionConvertor;
 import chess.domain.board.Board;
 import chess.domain.board.BoardFactory;
-import chess.dto.GameDto;
-import chess.dto.GamesDto;
-import chess.dto.StatusDto;
-import chess.dto.WinnerDto;
+import chess.dto.response.GameDto;
+import chess.dto.response.GamesDto;
+import chess.dto.response.StatusDto;
+import chess.dto.response.WinnerDto;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
